[refactor] 뷰, 뷰컨 파일 분리
